Don't force floating panel.
In case of StaysOnTop windows it looks like it is floating already.
But in case of manually setting floating here it has a strange bug
with non-activatable always on top tool window (picture-in-picture).

When first time shown it doesn't act as a stays-on-top window, if you
click on another app it is shown above the PiP. Only if you activate
the main window once more the PiP will become stays-on-top window.

This way it works from the first show.