Fix possible render glitch.
In some rare cases opening main menu was leading to some
widgets showing "over" the background darkening filling.
The glitch disappeared on the next repaint, so we trigger
this repaint explicitly.

The case it was reproducing with 100% rate: in TDesktop with
disabled interface animations after showing and hiding of bot
keyboard the message input field (widget with OpaquePaintEvent)
was painted on top of the backround dark filling below main menu.