Fix `Ui::Animations::Simple`'s `.start()` and `.change()` methods
Unlike the `.change()` method, which configures a running animation to
go to a different state, the `.start()` method should first set the
initial opacity value, then animate.

It is fixed by making sure the private `.prepare()` method always sets
`from`, rather than ignoring it when the animation is already running.